feat: add prompt generator for on the fly system prompt generation (#4060)
